Turned output buffering on.
This is a patch.  I screwed something up after the last import fix
(issue #10) commit and was getting an error telling me the system was
trying to redirect while it was still rendering output as soon as I
installed.  I couldn't actually figure it out, but stack exchange told
me this would fix it, and it did.